# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Article API",
    "description": "Article API to manage Admin users, Tenants, Articles, Tags and References.",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/api/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/categories": {
      "post": {
        "tags": ["Category"],
        "description": "Create new Category",
        "operationId": "createCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Category Parameters",
            "in": "body",
            "description": "Create new category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "403": {
            "description": "Category Duplication",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },      
      "get": {
        "tags": ["Category"],
        "description": "Get Category List",
        "operationId": "getCategories",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Category List",
            "schema": {
              "$ref": "#/definitions/GetCategory"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/categories/{_id}": {
      "patch": {
        "tags": ["Category"],
        "description": "Update Category",
        "operationId": "updateCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Category Parameters",
            "in": "body",
            "description": "Update Category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "Category Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Invalid Category Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": ["Category"],
        "description": "Delete Category",
        "operationId": "deleteCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "404": {
            "description": "Invalid Category Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/jurisdictions": {
      "post": {
        "tags": ["Jurisdiction"],
        "description": "Create new Jurisdiction",
        "operationId": "createJurisdiction",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Jurisdiction Parameters",
            "in": "body",
            "description": "Create new jurisdiction",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Jurisdiction"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "403": {
            "description": "Jurisdiction Duplication",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },      
      "get": {
        "tags": ["Jurisdiction"],
        "description": "Get Jurisdiction List",
        "operationId": "getJurisdictions",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Jurisdiction List",
            "schema": {
              "$ref": "#/definitions/GetJurisdiction"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/jurisdictions/{_id}": {
      "patch": {
        "tags": ["Jurisdiction"],
        "description": "Update Jurisdiction",
        "operationId": "updateJurisdiction",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Jurisdiction Parameters",
            "in": "body",
            "description": "Update Jurisdiction",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateJurisdiction"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "Jurisdiction Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Invalid Jurisdiction Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": ["Jurisdiction"],
        "description": "Delete Jurisdiction",
        "operationId": "deleteJurisdiction",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "404": {
            "description": "Invalid Jurisdiction Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateJurisdiction": {
      "properties": {
        "jurisdiction": {
          "required":true,
          "type": "string",
          "example": "jurisdiction"
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetJurisdiction": {
      "properties": {
        "jurisdiction": {
          "required":true,
          "type": "string",
          "example": "jurisdiction"
        }
      }
    },
    "Jurisdiction": {
      "properties": {
        "jurisdiction": {
          "required":true,
          "type": "string",
          "example": "jurisdiction"
        },
        "token": {
          "required":true,
          "type":"string",
          "example": "Token"
        } 
      }
    },
    "UpdateCategory": {
      "properties": {
        "category": {
          "required":true,
          "type": "string",
          "example": "category"
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetCategory": {
      "properties": {
        "category": {
          "required":true,
          "type": "string",
          "example": "category"
        }
      }
    },
    "Category": {
      "properties": {
        "category": {
          "required":true,
          "type": "string",
          "example": "category"
        },
        "token": {
          "required":true,
          "type":"string",
          "example": "Token"
        } 
      }
    },
...
  }
}